update app class and product class, remove test codes from index.html
