+++ server/public/index.html
@@ -19,29 +19,6 @@
 </head>
 <body>
   <div id="root">
-    <div class="main-container">
-      <!-- title -->
-      <div class="title p-3 mb-5 bg-dark text-white">
-        $Wicked Sales
-      </div>
-      <!-- products display -->
-      <div class="products-container">
-        <div class="row">
-            <!-- one card product -->
-            <div class="card col-md-3 mr-5 mb-5" style="width: 18rem;">
-              <img src="./images/shake-weight.jpg" class="card-img-top">
-              <div class="card-body" >
-                <h5 class="card-title">Product Name</h5>
-                <h6 class="card-subtitle mb-2 text-muted">$39.89</h6>
-                <p class="card-text">Some quick example text to build on the card title and make up the bulk of the card's content.
-                </p>
-              </div>
-            </div>
-
-          </div>
-        </div>
-    </div>
-    </div>
   </div>
   <script src="/main.js"></script>
 </body>
